Revert "Developer tools heading"
Right now, "Developer tools" is taking up double the size of the
second-largest section and triple the average size of all other
sections.

- Home: 3 squares
- CDK: 2 squares
- zkEVM: 4 squares
- PoS: 2 squares
- Miden: 3 squares
- Developer tools: 8 squares
- Learn: 3 squares

It's not appropriate design to have a miscellaneous section take up
triple the average navigation size of the main product sections.

Please accept this commit to revert the change.