Updating crossmint.md to include internal analytics
I made a few minor edits for us to keep track of analytics through utm_source and campaigns. 

This helps us understand the users that are being redirected to our docs, product, etc through the Polygon docs. Eventually, we will use the inferences drawn from the analytics to provide more catered content for the Polygon ecosystem and community