Update README.md
This document explains how to automatically write a smart contract using the OpenZeppelin Wizard. The resulting smart contract code can either be integrated with Remix by Clicking the **Open in Remix** button, or copied to a clipboard and pasted in the user's intended IDE.

## Getting started

# ./sol-merger \"./contracts/*.sol\" ./build

// Specifies that the source code is for a version
// of Solidity greater than 0.5.10
pragma solidity ^0.5.10;

// A contract is a collection of functions and data (its state)
// that resides at a specific address on the Ethereum blockchain.
contract HelloWorld {

    // The keyword "public" makes variables accessible from outside a contract
    // and creates a function that other contracts or SDKs can call to access the value
    string public message;

    // A special function only run during the creation of the contract
    constructor(string memory initMessage) public {
        // Takes a string value and stores the value in the memory data storage area,
        // setting `message` to that value
        message = initMessage;
    }

    // A publicly accessible function that takes a string as a parameter
    // and updates `message`
    function update(string memory newMessage) public {
        message = newMessage;
    }
}

Navigate to the [OpenZeppelin Wizard](https://wizard.openzeppelin.com) in your browser. First thing to notice is the **Solidity Wizard** and **Cairo Wizard** buttons.

One can choose any of the following tabs to begin creating an out-of-box smart contract code in either Solidity (for EVM chains) or Cairo (useful for Starknet). These are:

- **ERC20** for writing an ERC-20 token smart contract.
- **ERC721** for writing an NFT token smart contract.
- **ERC1155** for writing an ERC-1155 token smart contract.
- **Governor** for creating a DAO.
- **Custom** for writing a customized smart contract.

## Writing an NFT contract

For illustration purposes, we will be creating a NFT smart contract.

Suppose you wanted to create a `Mintable`, `Burnable` ERC721 token and specify an appropriate license for it.

1. Select the **ERC721** tab.

2. Give your NFT a name and a symbol by filling the `Name` and `Symbol` fields.

3. Use the check-boxes on the left to select features of your token.

- Put a tick on the `Mintable` check-box.
- Put a tick on the `Auto Increment Ids` check-box, this ensures uniqueness of each minted NFT.
- Put a tick on the `Burnable` check-box.
- Either leave the **default MIT license** or type the license of your choice.

  Notice that new lines of code are automatically written each time a feature is selected.

## Voila! Contract is ready

With the resulting lines of code, you now have the NFT token contract written in Solidity. As mentioned above, this source code can now be ported to an IDE of your choice or opened directly in Remix.

The below figure depicts the auto-written NFT smart contract code.

![The end-product NFT source code](../../img/zkEVM/zkv-end-product-nft-code.png)